Moar refact.
Push already instanciated modes instead of the leaving instanciation to
the mode manager (this was done to ensure client was passed to the mode,
but was ugly and error prone. This is python, just set the client attr
on the instance).

Also, fix various bugs left over from afca52d.