Fix pathfinding shenanigans.
Max recusrsion was reached when path to player was blocked (which froze
the game while the actor repeateadly tried to recompute the same path).

Fixed by aborting path computation as soon as we see target is not
reachable, which required improving the AI to try and deal with the
failure.

Works moslty okay. Still some trouble with ganging up around the player
standing at the entrance of a corridor, which is caused by the Dtra
distance computation ignoring diagonal movement. Fixing this is easy
enough, but may have other consequances, so we'll handle it (if at all)
in another commit.

Since we were messing with AI, took the chance to improve random move:
actors now check if the chosen tile is unoccupied before moving, which
avoids entities randomly killing each other offscreen.