Let's try a no higher cost for diagonals distaznce computation (see 723f10f1).
Behaviour seems okay for now, but we do take a performance hit
(manageable but for how long ?). We may or may nt decide to go back to
the old way in the future (higher diagonal cost could be a valid
gameplay decision), but let's spend some time with this version before
deciding.