Log a welcome message on game start (via a new GAME_STARTED event).
This involved fixing a bug causing events emitted during the very first
turn not to be logged. This was caused by using a temporary init game
mode before swithcing to the run one (start response was handled before
swithcung mode, and therefore ignored). Since that init mode was a stub,
the simple fix is to simply not to use it. Will need to be careful if
that first stage idea becomes relevant in the future, tho.